Data.Map.NonEmpty

Non-Empty Map type

data NEMap k a

A non-empty Map from keys k to values a. At least one key-value pair exists in an NEMap k v at all times.

Functions that take an NEMap can safely operate on it with the assumption that it has at least one key-value pair.

Functions that return an NEMap provide an assurance that the result has at least one key-value pair.

Data.Map.NonEmpty re-exports the API of Data.Map, faithfully reproducing asymptotics, typeclass constraints, and semantics. Functions that ensure that input and output maps are both non-empty (like insert) return NEMap, but functions that might potentially return an empty map (like delete) return a Map instead.

You can directly construct an NEMap with the API from Data.Map.NonEmpty; it's more or less the same as constructing a normal Map, except you don't have access to empty. There are also a few ways to construct an NEMap from a Map:

The nonEmptyMap smart constructor will convert a Map k a into a Maybe (NEMap k a), returning Nothing if the original Map was empty.

You can use the insertMap family of functions to insert a value into a Map to create a guarunteed NEMap.

You can convert an NEMap into a (non-empty) Map with toMap.

data Map k a where

A Map from keys k to values a.

Bundled Patterns

pattern NonEmptyMap :: NEMap k a -> Map k a
O(log n). The NonEmptyMap and EmptyMap patterns allow you to treat a Map as if it were either a NonEmptyMap n (where n is a NEMap) or an EmptyMap.

For example, you can pattern match on a Map:

myFunc :: Map K X -> Y
myFunc m = case m of
    NonEmptyMap n -> -- here, n is a NonEmptyMap, m is proven not empty.
    EmptyMap      -> -- here, m is empty.

Matching on NonEmptyMap n means that the original Map was not empty, and you have a verified-non-empty NEMap n to use.

Note that patching on this pattern is O(1). However, using the contents requires a O(log n) cost that is deferred until after the pattern is matched on (and is not incurred at all if the contents are never used).

pattern EmptyMap :: Map k a
O(1). The NonEmptyMap and EmptyMap patterns allow you to treat a Map as if it were either a NonEmptyMap n (where n is a NEMap) or an EmptyMap.

Matching on EmptyMap means that the original Map was empty.

See NonEmptyMap for more information.

nonEmptyMap :: Map k a -> Maybe (NEMap k a)

O(log n) Smart constructor for an NEMap from a Map. Returns Nothing if the Map was originally actually empty, and Just n with an NEMap, if the Map was not empty.

toMap :: NEMap k a -> Map k a

O(log n). Convert a non-empty map back into a normal possibly-empty map, for usage with functions that expect Map.

insertMap :: Ord k => k -> a -> Map k a -> NEMap k a

O(log n). Convert a Map into an NEMap by adding a key-value pair. If key is already present, will overwrite the original value.

See insertMapMin for a version that is constant-time if the new key is strictly smaller than all keys in the original map.

insertMapWith :: Ord k => (a -> a -> a) -> k -> a -> Map k a -> NEMap k a

O(log n). Convert a Map into an NEMap by adding a key-value pair. Uses a combining function with the new value as the first argument if the key is already present.

insertMapWithKey :: Ord k => (k -> a -> a -> a) -> k -> a -> Map k a -> NEMap k a

O(log n). Convert a Map into an NEMap by adding a key-value pair. Uses a combining function with the key and new value as the first and second arguments if the key is already present.

insertMapMin :: k -> a -> Map k a -> NEMap k a

O(1) Convert a Map into an NEMap by adding a key-value pair where the key is strictly less than all keys in the input map. The keys in the original map must all be strictly greater than the new key. The precondition is not checked.

insertMapMax :: k -> a -> Map k a -> NEMap k a

O(log n) Convert a Map into an NEMap by adding a key-value pair where the key is strictly greater than all keys in the input map. The keys in the original map must all be strictly less than the new key. The precondition is not checked.

While this has the same asymptotics as insertMap, it saves a constant factor for key comparison (so may be helpful if comparison is expensive) and also does not require an Ord instance for the key type.

Construction

empty :: Alternative f => f a

The identity of <|>

singleton :: k -> a -> NEMap k a

O(1). A map with a single element.

From Unordered Lists

fromList :: Ord k => NonEmpty (k, a) -> NEMap k a

O(n*log n). Build a non-empty map from a non-empty list of key/value pairs. See also fromAscList. If the list contains more than one value for the same key, the last value for the key is retained.

fromListWith :: Ord k => (a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

fromListWithKey :: Ord k => (k -> a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

From Ascending Lists

fromAscList :: Eq k => NonEmpty (k, a) -> NEMap k a

fromAscListWith :: Eq k => (a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

fromAscListWithKey :: Eq k => (k -> a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

fromDistinctAscList :: NonEmpty (k, a) -> NEMap k a

fromDescList :: Eq k => NonEmpty (k, a) -> NEMap k a

fromDescListWith :: Eq k => (a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

fromDescListWithKey :: Eq k => (k -> a -> a -> a) -> NonEmpty (k, a) -> NEMap k a

fromDistinctDescList :: NonEmpty (k, a) -> NEMap k a

Insertion

insert :: Ord k => k -> a -> NEMap k a -> NEMap k a

insertWith :: Ord k => (a -> a -> a) -> k -> a -> NEMap k a -> NEMap k a

O(log n). Insert with a function, combining new value and old value. insertWith f key value mp will insert the pair (key, value) into mp if key does not exist in the map. If the key does exist, the function will insert the pair (key, f new_value old_value).

insertWithKey :: Ord k => (k -> a -> a -> a) -> k -> a -> NEMap k a -> NEMap k a

insertLookupWithKey :: Ord k => (k -> a -> a -> a) -> k -> a -> NEMap k a -> (Maybe a, NEMap k a)

Deletion/Update

delete :: Ord k => k -> NEMap k a -> Map k a

adjust :: Ord k => (a -> a) -> k -> NEMap k a -> NEMap k a

adjustWithKey :: Ord k => (k -> a -> a) -> k -> NEMap k a -> NEMap k a

update :: Ord k => (a -> Maybe a) -> k -> NEMap k a -> Map k a

updateWithKey :: Ord k => (k -> a -> Maybe a) -> k -> NEMap k a -> Map k a

updateLookupWithKey :: Ord k => (k -> a -> Maybe a) -> k -> NEMap k a -> (Maybe a, Map k a)

alter :: Ord k => (Maybe a -> Maybe a) -> k -> NEMap k a -> Map k a

alterF :: (Ord k, Functor f) => (Maybe a -> f (Maybe a)) -> k -> NEMap k a -> f (Map k a)

alter' :: Ord k => (Maybe a -> a) -> k -> NEMap k a -> NEMap k a

Variant of alter that disallows deletion. Allows us to guarantee that the result is also a non-empty Map.

alterF' :: (Ord k, Functor f) => (Maybe a -> f a) -> k -> NEMap k a -> f (NEMap k a)

Variant of alterF that disallows deletion. Allows us to guarantee that the result is also a non-empty Map.

TODO: RULES

Query

Lookup

lookup :: Ord k => k -> NEMap k a -> Maybe a

O(log n). Find largest key smaller than the given one and return the corresponding (key, value) pair.

(!?) :: Ord k => NEMap k a -> k -> Maybe a

O(log n). Find the value at a key. Returns Nothing when the element can not be found.

(!) :: Ord k => NEMap k a -> k -> a

O(log n). Find the value at a key. Calls error when the element can not be found.

findWithDefault :: Ord k => a -> k -> NEMap k a -> a

O(log n). The expression (findWithDefault def k map) returns the value at key k or returns default value def when the key is not in the map.

member :: Ord k => k -> NEMap k a -> Bool

O(log n). Is the key a member of the map? See also notMember.

notMember :: Ord k => k -> NEMap k a -> Bool

O(log n). Is the key not a member of the map? See also member.

lookupLT :: Ord k => k -> NEMap k a -> Maybe (k, a)

O(log n). Find largest key smaller than the given one and return the corresponding (key, value) pair.

lookupGT :: Ord k => k -> NEMap k a -> Maybe (k, a)

O(log n). Find smallest key greater than the given one and return the corresponding (key, value) pair.

lookupLE :: Ord k => k -> NEMap k a -> Maybe (k, a)

O(log n). Find largest key smaller or equal to the given one and return the corresponding (key, value) pair.

lookupGE :: Ord k => k -> NEMap k a -> Maybe (k, a)

O(log n). Find smallest key greater or equal to the given one and return the corresponding (key, value) pair.

Size

null :: Foldable t => t a -> Bool

Test whether the structure is empty. The default implementation is optimized for structures that are similar to cons-lists, because there is no general way to do better.

size :: NEMap k a -> Int

O(1). The number of elements in the map.

Combine

Union

union :: Ord k => NEMap k a -> NEMap k a -> NEMap k a

O(m*log(n/m + 1)), m <= n. The expression (union t1 t2) takes the left-biased union of t1 and t2. It prefers t1 when duplicate keys are encountered, i.e. (union == unionWith const).

unionWith :: Ord k => (a -> a -> a) -> NEMap k a -> NEMap k a -> NEMap k a

O(m*log(n/m + 1)), m <= n. Union with a combining function.

unionWithKey :: Ord k => (k -> a -> a -> a) -> NEMap k a -> NEMap k a -> NEMap k a

O(m*log(n/m + 1)), m <= n. Union with a combining function, given the matching key.

unions :: (Foldable1 f, Ord k) => f (NEMap k a) -> NEMap k a

The left-biased union of a list of maps.

unionsWith :: (Foldable1 f, Ord k) => (a -> a -> a) -> f (NEMap k a) -> NEMap k a

The union of a list of maps, with a combining operation.

Difference

difference :: Ord k => NEMap k a -> NEMap k b -> Map k a

O(m*log(n/m + 1)), m <= n. Difference of two maps. Return elements of the first map not existing in the second map.

(\\) :: Ord k => NEMap k a -> NEMap k b -> Map k a

Same as difference.

differenceWith :: Ord k => (a -> b -> Maybe a) -> NEMap k a -> NEMap k b -> Map k a

O(n+m). Difference with a combining function. When two equal keys are encountered, the combining function is applied to the values of these keys. If it returns Nothing, the element is discarded (proper set difference). If it returns (Just y), the element is updated with a new value y.

differenceWithKey :: Ord k => (k -> a -> b -> Maybe a) -> NEMap k a -> NEMap k b -> Map k a

O(n+m). Difference with a combining function. When two equal keys are encountered, the combining function is applied to the key and both values. If it returns Nothing, the element is discarded (proper set difference). If it returns (Just y), the element is updated with a new value y.

intersection :: Ord k => NEMap k a -> NEMap k b -> Map k a

O(m*log(n/m + 1)), m <= n. Intersection of two maps. Return data in the first map for the keys existing in both maps. (intersection m1 m2 == intersectionWith const m1 m2).

intersectionWith :: Ord k => (a -> b -> c) -> NEMap k a -> NEMap k b -> Map k c

O(m*log(n/m + 1)), m <= n. Intersection with a combining function.

intersectionWithKey :: Ord k => (k -> a -> b -> c) -> NEMap k a -> NEMap k b -> Map k c

O(m*log(n/m + 1)), m <= n. Intersection with a combining function.

Traversal

Map

map :: (a -> b) -> NEMap k a -> NEMap k b

mapWithKey :: (k -> a -> b) -> NEMap k a -> NEMap k b

O(n). Map a function over all values in the map.

traverseWithKey1 :: Apply t => (k -> a -> t b) -> NEMap k a -> t (NEMap k b)

O(n). traverseWithKey1 f m == fromNonEmpty $ traverse1 ((k, v) -> (,) k $ f k v) (toList m)

That is, behaves exactly like a regular traverse1 except that the traversing function also has access to the key associated with a value.

traverseWithKey :: Applicative t => (k -> a -> t b) -> NEMap k a -> t (NEMap k b)

O(n). traverseWithKey f m == fromNonEmpty $ traverse ((k, v) -> (,) k $ f k v) (toList m) That is, behaves exactly like a regular traverse except that the traversing function also has access to the key associated with a value.

Use traverseWithKey1 whenever possible (if your Applicative also has Apply instance). This version is provided only for types that do not have Apply instance, since Apply is not at the moment (and might not ever be) an official superclass of Applicative.

traverseWithKey f = unwrapApplicative . traverseWithKey1 (\k -> WrapApplicative . f k)

traverseMaybeWithKey :: Apply t => (k -> a -> t (Maybe b)) -> NEMap k a -> t (Map k b)

mapAccum :: (a -> b -> (a, c)) -> a -> NEMap k b -> (a, NEMap k c)

mapAccumWithKey :: (a -> k -> b -> (a, c)) -> a -> NEMap k b -> (a, NEMap k c)

mapAccumRWithKey :: (a -> k -> b -> (a, c)) -> a -> NEMap k b -> (a, NEMap k c)

mapKeys :: Ord k2 => (k1 -> k2) -> NEMap k1 a -> NEMap k2 a

mapKeysWith :: Ord k2 => (a -> a -> a) -> (k1 -> k2) -> NEMap k1 a -> NEMap k2 a

mapKeysMonotonic :: (k1 -> k2) -> NEMap k1 a -> NEMap k2 a

Folds

foldr :: (a -> b -> b) -> b -> NEMap k a -> b

O(n). Fold the values in the map using the given right-associative binary operator, such that foldr f z == foldr f z . elems.

foldl :: (a -> b -> a) -> a -> NEMap k b -> a

O(n). Fold the values in the map using the given left-associative binary operator, such that foldl f z == foldl f z . elems.

foldrWithKey :: (k -> a -> b -> b) -> b -> NEMap k a -> b

O(n). Fold the keys and values in the map using the given right-associative binary operator, such that foldrWithKey f z == foldr (uncurry f) z . toAscList.

For example,

keysList map = foldrWithKey (\k x ks -> k:ks) [] map

foldlWithKey :: (a -> k -> b -> a) -> a -> NEMap k b -> a

O(n). Fold the keys and values in the map using the given left-associative binary operator, such that foldlWithKey f z == foldl (\z' (kx, x) -> f z' kx x) z . toAscList.

For example,

keysList = reverse . foldlWithKey (\ks k x -> k:ks) []

foldMapWithKey :: Semigroup m => (k -> a -> m) -> NEMap k a -> m

O(n). Fold the keys and values in the map using the given semigroup, such that

foldMapWithKey f = fold1 . mapWithKey f

This can be an asymptotically faster than foldrWithKey or foldlWithKey for some monoids.

Strict folds

foldr' :: (a -> b -> b) -> b -> NEMap k a -> b

O(n). A strict version of foldr. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldr1' :: (a -> a -> a) -> NEMap k a -> a

O(n). A strict version of foldr1. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldl' :: (a -> b -> a) -> a -> NEMap k b -> a

O(n). A strict version of foldl. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldl1 :: (a -> a -> a) -> NEMap k a -> a

O(n). A version of foldl that uses the value at the minimal key in the map as the starting value.

Note that, unlike foldl1 for Map, this function is total if the input function is total.

foldl1' :: (a -> a -> a) -> NEMap k a -> a

O(n). A strict version of foldl1. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldrWithKey' :: (k -> a -> b -> b) -> b -> NEMap k a -> b

O(n). A strict version of foldrWithKey. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldlWithKey' :: (a -> k -> b -> a) -> a -> NEMap k b -> a

O(n). A strict version of foldlWithKey. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

Conversion

elems :: NEMap k a -> NonEmpty a

O(n). Return all elements of the map in the ascending order of their keys.

keys :: NEMap k a -> NonEmpty k

O(n). Return all keys of the map in ascending order.

assocs :: NEMap k a -> NonEmpty (k, a)

O(n). An alias for toAscList. Return all key/value pairs in the map in ascending key order.

Lists

toList :: NEMap k a -> NonEmpty (k, a)

O(n). Convert the map to a non-empty list of key/value pairs.

Ordered lists

toAscList :: NEMap k a -> NonEmpty (k, a)

O(n). Convert the map to a list of key/value pairs where the keys are in ascending order.

toDescList :: NEMap k a -> NonEmpty (k, a)

O(n). Convert the map to a list of key/value pairs where the keys are in descending order.

Filter

filter :: (a -> Bool) -> NEMap k a -> Map k a

filterWithKey :: (k -> a -> Bool) -> NEMap k a -> Map k a

restrictKeys :: Ord k => NEMap k a -> Set k -> Map k a

O(m*log(n/m + 1)), m <= n. Restrict an NEMap to only those keys found in a Set.

m `restrictKeys` s = filterWithKey (k _ -> k `member` s) m
m `restrictKeys` s = m `intersect` fromSet (const ()) s

withoutKeys :: Ord k => NEMap k a -> Set k -> Map k a

O(m*log(n/m + 1)), m <= n. Remove all keys in a Set from an NEMap.

m `withoutKeys` s = filterWithKey (k _ -> k `notMember` s) m
m `withoutKeys` s = m `difference` fromSet (const ()) s

partition :: (a -> Bool) -> NEMap k a -> (Map k a, Map k a)

partitionWithKey :: (k -> a -> Bool) -> NEMap k a -> (Map k a, Map k a)

takeWhileAntitone :: (k -> Bool) -> NEMap k a -> Map k a

dropWhileAntitone :: (k -> Bool) -> NEMap k a -> Map k a

spanAntitone :: (k -> Bool) -> NEMap k a -> (Map k a, Map k a)

mapMaybe :: (a -> Maybe b) -> NEMap k a -> Map k b

mapMaybeWithKey :: (k -> a -> Maybe b) -> NEMap k a -> Map k b

mapEither :: (a -> Either b c) -> NEMap k a -> (Map k b, Map k c)

mapEitherWithKey :: (k -> a -> Either b c) -> NEMap k a -> (Map k b, Map k c)

split :: Ord k => k -> NEMap k a -> (Map k a, Map k a)

splitLookup :: Ord k => k -> NEMap k a -> (Map k a, Maybe a, Map k a)

splitRoot :: NEMap k a -> NonEmpty (NEMap k a)

Submap

isSubmapOf :: (Ord k, Eq a) => NEMap k a -> NEMap k a -> Bool

isSubmapOfBy :: Ord k => (a -> b -> Bool) -> NEMap k a -> NEMap k b -> Bool

isProperSubmapOf :: (Ord k, Eq a) => NEMap k a -> NEMap k a -> Bool

isProperSubmapOfBy :: Ord k => (a -> b -> Bool) -> NEMap k a -> NEMap k b -> Bool

lookupIndex :: Ord k => k -> NEMap k a -> Maybe Int

findIndex :: Ord k => k -> NEMap k a -> Int

elemAt :: Int -> NEMap k a -> (k, a)

adjustAt :: (k -> a -> a) -> Int -> NEMap k a -> NEMap k a

Variant of updateAt that disallows deletion. Allows us to guarantee that the result is also a non-empty Map.

updateAt :: (k -> a -> Maybe a) -> Int -> NEMap k a -> Map k a

deleteAt :: Int -> NEMap k a -> Map k a

take :: Int -> NEMap k a -> Map k a

drop :: Int -> NEMap k a -> Map k a

splitAt :: Int -> NEMap k a -> (Map k a, Map k a)

Min/Max

findMin :: NEMap k a -> (k, a)

findMax :: NEMap k a -> (k, a)

deleteMin :: NEMap k a -> Map k a

deleteMax :: NEMap k a -> Map k a

deleteFindMin :: NEMap k a -> ((k, a), Map k a)

deleteFindMax :: NEMap k a -> ((k, a), Map k a)

updateMin :: (a -> Maybe a) -> NEMap k a -> Map k a

updateMax :: (a -> Maybe a) -> NEMap k a -> Map k a

updateMinWithKey :: (k -> a -> Maybe a) -> NEMap k a -> Map k a

updateMaxWithKey :: (k -> a -> Maybe a) -> NEMap k a -> Map k a

minView :: NEMap k a -> (a, Map k a)

maxView :: NEMap k a -> (a, Map k a)

Debugging

valid :: Ord k => NEMap k a -> Bool

O(n). Test if the internal map structure is valid.