Data.Sequence.NonEmpty

Finite sequences

data NESeq a

Constructors

a :<|| !(Seq a)

Bundled Patterns

pattern (:||>) :: Seq a -> a -> NESeq a

Conversions between empty and non-empty sequences

nonEmptySeq :: Seq a -> Maybe (NESeq a)

toSeq :: NESeq a -> Seq a

withNonEmpty :: r -> (NESeq a -> r) -> Seq a -> r

unsafeFromSeq :: Seq a -> NESeq a

Construction

singleton :: a -> NESeq a

(<|) :: a -> NESeq a -> NESeq a

(|>) :: NESeq a -> a -> NESeq a

(><) :: NESeq a -> NESeq a -> NESeq a

fromList :: NonEmpty a -> NESeq a

Repetition

replicate :: Int -> a -> NESeq a

replicateA :: Applicative f => Int -> f a -> f (NESeq a)

replicateA1 :: Apply f => Int -> f a -> f (NESeq a)

replicateM :: Applicative m => Int -> m a -> m (NESeq a)

Iterative construction

iterateN :: Int -> (a -> a) -> a -> NESeq a

unfoldr :: (a -> (b, Maybe a)) -> a -> NESeq b

unfoldl :: (a -> (b, Maybe a)) -> a -> NESeq b

Deconstruction

Additional functions for deconstructing sequences are available via the Foldable instance of Seq.

Queries

length :: NESeq a -> Int

Scans

scanl :: (a -> b -> a) -> a -> NESeq b -> NESeq a

scanl1 :: (a -> a -> a) -> NESeq a -> NESeq a

scanr :: (a -> b -> b) -> b -> NESeq a -> NESeq b

scanr1 :: (a -> a -> a) -> NESeq a -> NESeq a

Sublists

tails :: NESeq a -> NESeq (NESeq a)

inits :: NESeq a -> NESeq (NESeq a)

chunksOf :: Int -> NESeq a -> NESeq (NESeq a)

Sequential searches

takeWhileL :: (a -> Bool) -> NESeq a -> Seq a

takeWhileR :: (a -> Bool) -> NESeq a -> Seq a

dropWhileL :: (a -> Bool) -> NESeq a -> Seq a

dropWhileR :: (a -> Bool) -> NESeq a -> Seq a

spanl :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

spanr :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

breakl :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

breakr :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

partition :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

filter :: (a -> Bool) -> NESeq a -> Seq a

splitAt :: Int -> NESeq a -> These (NESeq a) (NESeq a)

Indexing with predicates

These functions perform sequential searches from the left or right ends of the sequence returning indices of matching elements. , elemIndexL , elemIndicesL , elemIndexR , elemIndicesR , findIndexL , findIndicesL , findIndexR , findIndicesR -- * Folds -- | General folds are available via the Foldable instance of Seq. , foldMapWithIndex , foldlWithIndex , foldrWithIndex -- * Transformations , mapWithIndex , traverseWithIndex , reverse , intersperse ** Zips and unzip

zip :: NESeq a -> NESeq b -> NESeq (a, b)

zipWith :: (a -> b -> c) -> NESeq a -> NESeq b -> NESeq c

zip3 :: NESeq a -> NESeq b -> NESeq c -> NESeq (a, b, c)

zipWith3 :: (a -> b -> c -> d) -> NESeq a -> NESeq b -> NESeq c -> NESeq d

zip4 :: NESeq a -> NESeq b -> NESeq c -> NESeq d -> NESeq (a, b, c, d)

zipWith4 :: (a -> b -> c -> d -> e) -> NESeq a -> NESeq b -> NESeq c -> NESeq d -> NESeq e

unzip :: NESeq (a, b) -> (NESeq a, NESeq b)

unzipWith :: (a -> (b, c)) -> NESeq a -> (NESeq b, NESeq c)