Data.Set.NonEmpty

Non-Empty Set Type

data NESet a

A non-empty set of values a. At least one value exists in an NESet a at all times.

Functions that take an NESet can safely operate on it with the assumption that it has at least one item.

Functions that return an NESet provide an assurance that the result has at least one item.

Data.Set.NonEmpty re-exports the API of Data.Set, faithfully reproducing asymptotics, typeclass constraints, and semantics. Functions that ensure that input and output sets are both non-empty (like insert) return NESet, but functions that might potentially return an empty map (like delete) return a Set instead.

You can directly construct an NESet with the API from Data.Set.NonEmpty; it's more or less the same as constructing a normal Set, except you don't have access to empty. There are also a few ways to construct an NESet from a Set:

The nonEmptySet smart constructor will convert a Set a into a Maybe (NESet a), returning Nothing if the original Set was empty.

You can use the insertSet family of functions to insert a value into a Set to create a guarunteed NESet.

You can use the IsNonEmpty and IsEmpty patterns to "pattern match" on a Set to reveal it as either containing a NESet or an empty map.

withNESet offers a continuation-based interface for deconstructing a Set and treating it as if it were an NESet.

You can convert an NESet into a (possibly empty) Set with toSet or IsNonEmpty, essentially "obscuring" the non-empty property from the type.

Conversions between empty and non-empty maps

pattern IsNonEmpty :: NESet a -> Set a

pattern IsEmpty :: Set a

nonEmptySet :: Set a -> Maybe (NESet a)

O(log n). Smart constructor for an NESet from a Set. Returns Nothing if the Set was originally actually empty, and Just n with an NESet, if the Set was not empty.

nonEmptySet and maybe 'Data.Set..empty' toSet form an isomorphism: they are perfect structure-preserving inverses of eachother.

See IsNonEmpty for a pattern synonym that lets you "match on" the possiblity of a Set being an NESet.

nonEmptySet (Data.Set.fromList [3,5]) == fromList (3:|[5])

toSet :: NESet a -> Set a

O(log n). Convert a non-empty set back into a normal possibly-empty map, for usage with functions that expect Set.

Can be thought of as "obscuring" the non-emptiness of the set in its type. See the IsNotEmpty pattern.

nonEmptySet and maybe empty toSet form an isomorphism: they are perfect structure-preserving inverses of eachother.

toSet (fromList ((3,"a") :| [(5,"b")])) == Data.Set.fromList [(3,"a"), (5,"b")]

withNESet

Arguments

:: r
value to return if set is empty

-> (NESet a -> r)
function to apply if set is not empty

-> Set a

-> r

insertSet :: Ord a => a -> Set a -> NESet a

insertSetMin :: a -> Set a -> NESet a

insertSetMax :: a -> Set a -> NESet a

unsafeFromSet :: Set a -> NESet a

O(log n). Unsafe version of nonEmptySet. Coerces a Set into an NESet, but is undefined (throws a runtime exception when evaluation is attempted) for an empty Set.

Construction

singleton :: a -> NESet a

O(1). Create a singleton set.

fromList :: Ord a => NonEmpty a -> NESet a

O(n*log n). Create a set from a list of elements.

fromAscList :: Eq a => NonEmpty a -> NESet a

fromDescList :: Eq a => NonEmpty a -> NESet a

fromDistinctAscList :: NonEmpty a -> NESet a

fromDistinctDescList :: NonEmpty a -> NESet a

powerSet :: forall a. NESet a -> NESet (NESet a)

insert :: Ord a => a -> NESet a -> NESet a

O(log n). Insert an element in a set. If the set already contains an element equal to the given value, it is replaced with the new value.

delete :: Ord a => a -> NESet a -> Set a

member :: Ord a => a -> NESet a -> Bool

notMember :: Ord a => a -> NESet a -> Bool

lookupLT :: Ord a => a -> NESet a -> Maybe a

lookupGT :: Ord a => a -> NESet a -> Maybe a

lookupLE :: Ord a => a -> NESet a -> Maybe a

lookupGE :: Ord a => a -> NESet a -> Maybe a

size :: NESet a -> Int

O(1). The number of elements in the set. Guaranteed to be greater than zero.

isSubsetOf :: Ord a => NESet a -> NESet a -> Bool

isProperSubsetOf :: Ord a => NESet a -> NESet a -> Bool

disjoint :: Ord a => NESet a -> NESet a -> Bool

union :: Ord a => NESet a -> NESet a -> NESet a

O(m*log(n/m + 1)), m <= n. The union of two sets, preferring the first set when equal elements are encountered.

unions :: (Foldable1 f, Ord a) => f (NESet a) -> NESet a

The union of a non-empty list of sets

difference :: Ord a => NESet a -> NESet a -> Set a

(\\) :: Ord a => NESet a -> NESet a -> Set a

Same as difference.

intersection :: Ord a => NESet a -> NESet a -> Set a

cartesianProduct :: NESet a -> NESet b -> NESet (a, b)

disjointUnion :: NESet a -> NESet b -> NESet (Either a b)

filter :: (a -> Bool) -> NESet a -> Set a

takeWhileAntitone :: (a -> Bool) -> NESet a -> Set a

dropWhileAntitone :: (a -> Bool) -> NESet a -> Set a

spanAntitone :: (a -> Bool) -> NESet a -> These (NESet a) (NESet a)

partition :: (a -> Bool) -> NESet a -> These (NESet a) (NESet a)

split :: Ord a => a -> NESet a -> Maybe (These (NESet a) (NESet a))

splitMember :: Ord a => a -> NESet a -> (Bool, Maybe (These (NESet a) (NESet a)))

splitRoot :: NESet a -> NonEmpty (NESet a)

lookupIndex :: Ord a => a -> NESet a -> Maybe Int

findIndex :: Ord a => a -> NESet a -> Int

elemAt :: Int -> NESet a -> a

deleteAt :: Int -> NESet a -> Set a

take :: Int -> NESet a -> Set a

drop :: Int -> NESet a -> Set a

splitAt :: Int -> NESet a -> These (NESet a) (NESet a)

map :: Ord b => (a -> b) -> NESet a -> NESet b

mapMonotonic :: (a -> b) -> NESet a -> NESet b

Folds

foldr :: (a -> b -> b) -> b -> NESet a -> b

O(n). Fold the elements in the set using the given right-associative binary operator, such that foldr f z == foldr f z . toAscList.

For example,

elemsList set = foldr (:) [] set

foldl :: (a -> b -> a) -> a -> NESet b -> a

O(n). Fold the elements in the set using the given left-associative binary operator, such that foldl f z == foldl f z . toAscList.

For example,

descElemsList set = foldl (flip (:)) [] set

foldr1 :: Foldable t => (a -> a -> a) -> t a -> a

A variant of foldr that has no base case, and thus may only be applied to non-empty structures.

foldr1 f = foldr1 f . toList

foldl1 :: Foldable t => (a -> a -> a) -> t a -> a

A variant of foldl that has no base case, and thus may only be applied to non-empty structures.

foldl1 f = foldl1 f . toList

Strict folds

foldr' :: (a -> b -> b) -> b -> NESet a -> b

O(n). A strict version of foldr. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldl' :: (a -> b -> a) -> a -> NESet b -> a

O(n). A strict version of foldl. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldr1' :: (a -> a -> a) -> NESet a -> a

foldl1' :: (a -> a -> a) -> NESet a -> a

findMin :: NESet a -> a

findMax :: NESet a -> a

deleteMin :: NESet a -> Set a

deleteMax :: NESet a -> Set a

deleteFindMin :: NESet a -> (a, Set a)

deleteFindMax :: NESet a -> (a, Set a)

elems :: NESet a -> NonEmpty a

toList :: NESet a -> NonEmpty a

O(n). Convert the set to a non-empty list of elements.

toAscList :: NESet a -> NonEmpty a

toDescList :: NESet a -> NonEmpty a

Debugging

valid :: Ord a => NESet a -> Bool

O(n). Test if the internal set structure is valid.