Data.Sequence.NonEmpty

Finite sequences

data NESeq a where infixr 5

Bundled Patterns

pattern (:<||) :: a -> Seq a -> NESeq a infixr 5

pattern (:||>) :: Seq a -> a -> NESeq a infixl 5

Conversions between empty and non-empty sequences

pattern IsNonEmpty :: NESeq a -> Seq a

O(1). The IsNonEmpty and IsEmpty patterns allow you to treat a Seq as if it were either a IsNonEmpty n (where n is a NESeq) or an IsEmpty.

For example, you can pattern match on a Seq:

safeHead :: Seq Int -> Int
safeHead (IsNonEmpty (x :<|| _))  = x  -- here, user provided a non-empty sequence, and n is the NESeq
safeHead IsEmpty                  = 0  -- here the user provided an empty sequence

Matching on IsNonEmpty n means that the original Seq was not empty, and you have a verified-non-empty NESeq n to use.

A case statement handling both IsNonEmpty and IsEmpty provides complete coverage.

This is a bidirectional pattern, so you can use IsNonEmpty to convert a NESeq back into a Seq, obscuring its non-emptiness (see toSeq).

pattern IsEmpty :: Seq a

O(1). The IsNonEmpty and IsEmpty patterns allow you to treat a Seq as if it were either a IsNonEmpty n (where n is a NESeq) or an IsEmpty.

Matching on IsEmpty means that the original Seq was empty.

A case statement handling both IsNonEmpty and IsEmpty provides complete coverage.

This is a bidirectional pattern, so you can use IsEmpty as an expression, and it will be interpreted as empty.

See IsNonEmpty for more information.

nonEmptySeq :: Seq a -> Maybe (NESeq a)

toSeq :: NESeq a -> Seq a

withNonEmpty :: r -> (NESeq a -> r) -> Seq a -> r

unsafeFromSeq :: Seq a -> NESeq a

insertSeqAt :: Int -> a -> Seq a -> NESeq a

Construction

singleton :: a -> NESeq a

(<|) :: a -> NESeq a -> NESeq a infixr 5

(|>) :: NESeq a -> a -> NESeq a infixl 5

(><) :: NESeq a -> NESeq a -> NESeq a infixr 5

(|><) :: NESeq a -> Seq a -> NESeq a infixr 5

(><|) :: Seq a -> NESeq a -> NESeq a infixr 5

fromList :: NonEmpty a -> NESeq a

fromFunction :: Int -> (Int -> a) -> NESeq a

Repetition

replicate :: Int -> a -> NESeq a

replicateA :: Applicative f => Int -> f a -> f (NESeq a)

replicateA1 :: Apply f => Int -> f a -> f (NESeq a)

replicateM :: Applicative m => Int -> m a -> m (NESeq a)

cycleTaking :: Int -> NESeq a -> NESeq a

Iterative construction

iterateN :: Int -> (a -> a) -> a -> NESeq a

unfoldr :: (a -> (b, Maybe a)) -> a -> NESeq b

unfoldl :: (a -> (b, Maybe a)) -> a -> NESeq b

Deconstruction

Additional functions for deconstructing sequences are available via the Foldable instance of Seq.

toList :: NESeq a -> NonEmpty a

Queries

length :: NESeq a -> Int

Scans

scanl :: (a -> b -> a) -> a -> NESeq b -> NESeq a

scanl1 :: (a -> a -> a) -> NESeq a -> NESeq a

scanr :: (a -> b -> b) -> b -> NESeq a -> NESeq b

scanr1 :: (a -> a -> a) -> NESeq a -> NESeq a

Sublists

tails :: NESeq a -> NESeq (NESeq a)

inits :: NESeq a -> NESeq (NESeq a)

chunksOf :: Int -> NESeq a -> NESeq (NESeq a)

Sequential searches

takeWhileL :: (a -> Bool) -> NESeq a -> Seq a

takeWhileR :: (a -> Bool) -> NESeq a -> Seq a

dropWhileL :: (a -> Bool) -> NESeq a -> Seq a

dropWhileR :: (a -> Bool) -> NESeq a -> Seq a

spanl :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

spanr :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

breakl :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

breakr :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

partition :: (a -> Bool) -> NESeq a -> These (NESeq a) (NESeq a)

filter :: (a -> Bool) -> NESeq a -> Seq a

Sorting

sort :: Ord a => NESeq a -> NESeq a

sortBy :: (a -> a -> Ordering) -> NESeq a -> NESeq a

sortOn :: Ord b => (a -> b) -> NESeq a -> NESeq a

unstableSort :: Ord a => NESeq a -> NESeq a

unstableSortBy :: (a -> a -> Ordering) -> NESeq a -> NESeq a

unstableSortOn :: Ord b => (a -> b) -> NESeq a -> NESeq a

Indexing

lookup :: Int -> NESeq a -> Maybe a

(!?) :: NESeq a -> Int -> Maybe a

index :: NESeq a -> Int -> a

adjust :: (a -> a) -> Int -> NESeq a -> NESeq a

adjust' :: (a -> a) -> Int -> NESeq a -> NESeq a

update :: Int -> a -> NESeq a -> NESeq a

take :: Int -> NESeq a -> Seq a

drop :: Int -> NESeq a -> Seq a

insertAt :: Int -> a -> NESeq a -> NESeq a

deleteAt :: Int -> NESeq a -> Seq a

splitAt :: Int -> NESeq a -> These (NESeq a) (NESeq a)

Indexing with predicates

These functions perform sequential searches from the left or right ends of the sequence returning indices of matching elements.

elemIndexL :: Eq a => a -> NESeq a -> Maybe Int

elemIndicesL :: Eq a => a -> NESeq a -> [Int]

elemIndexR :: Eq a => a -> NESeq a -> Maybe Int

elemIndicesR :: Eq a => a -> NESeq a -> [Int]

findIndexL :: (a -> Bool) -> NESeq a -> Maybe Int

findIndicesL :: (a -> Bool) -> NESeq a -> [Int]

findIndexR :: (a -> Bool) -> NESeq a -> Maybe Int

findIndicesR :: (a -> Bool) -> NESeq a -> [Int]

Folds

General folds are available via the Foldable instance of Seq.

foldMapWithIndex :: Semigroup m => (Int -> a -> m) -> NESeq a -> m

foldlWithIndex :: (b -> Int -> a -> b) -> b -> NESeq a -> b

foldrWithIndex :: (Int -> a -> b -> b) -> b -> NESeq a -> b

Transformations

mapWithIndex :: (Int -> a -> b) -> NESeq a -> NESeq b

traverseWithIndex :: Applicative f => (Int -> a -> f b) -> NESeq a -> f (NESeq b)

traverseWithIndex1 :: Apply f => (Int -> a -> f b) -> NESeq a -> f (NESeq b)

reverse :: NESeq a -> NESeq a

intersperse :: a -> NESeq a -> NESeq a

Zips and unzip

zip :: NESeq a -> NESeq b -> NESeq (a, b)

zipWith :: (a -> b -> c) -> NESeq a -> NESeq b -> NESeq c

zip3 :: NESeq a -> NESeq b -> NESeq c -> NESeq (a, b, c)

zipWith3 :: (a -> b -> c -> d) -> NESeq a -> NESeq b -> NESeq c -> NESeq d

zip4 :: NESeq a -> NESeq b -> NESeq c -> NESeq d -> NESeq (a, b, c, d)

zipWith4 :: (a -> b -> c -> d -> e) -> NESeq a -> NESeq b -> NESeq c -> NESeq d -> NESeq e

unzip :: NESeq (a, b) -> (NESeq a, NESeq b)

unzipWith :: (a -> (b, c)) -> NESeq a -> (NESeq b, NESeq c)