Data.Set.NonEmpty.Internal

Description

Unsafe internal-use functions used in the implementation of Data.Set.NonEmpty. These functions can potentially be used to break the abstraction of NESet and produce unsound sets, so be wary!

Documentation

data NESet a

A non-empty (by construction) set of values a. At least one value exists in an NESet a at all times.

Functions that take an NESet can safely operate on it with the assumption that it has at least one item.

Functions that return an NESet provide an assurance that the result has at least one item.

Data.Set.NonEmpty re-exports the API of Data.Set, faithfully reproducing asymptotics, typeclass constraints, and semantics. Functions that ensure that input and output sets are both non-empty (like insert) return NESet, but functions that might potentially return an empty map (like delete) return a Set instead.

You can directly construct an NESet with the API from Data.Set.NonEmpty; it's more or less the same as constructing a normal Set, except you don't have access to empty. There are also a few ways to construct an NESet from a Set:

The nonEmptySet smart constructor will convert a Set a into a Maybe (NESet a), returning Nothing if the original Set was empty.

You can use the insertSet family of functions to insert a value into a Set to create a guaranteed NESet.

You can use the IsNonEmpty and IsEmpty patterns to "pattern match" on a Set to reveal it as either containing a NESet or an empty map.

withNonEmpty offers a continuation-based interface for deconstructing a Set and treating it as if it were an NESet.

You can convert an NESet into a Set with toSet or IsNonEmpty, essentially "obscuring" the non-empty property from the type.

Constructors

NESet

Fields

nesV0 :: !a

invariant: must be smaller than smallest value in set

nesSet :: !(Set a)

nonEmptySet :: Set a -> Maybe (NESet a)

O(log n). Smart constructor for an NESet from a Set. Returns Nothing if the Set was originally actually empty, and Just n with an NESet, if the Set was not empty.

nonEmptySet and maybe empty toSet form an isomorphism: they are perfect structure-preserving inverses of eachother.

See IsNonEmpty for a pattern synonym that lets you "match on" the possiblity of a Set being an NESet.

nonEmptySet (Data.Set.fromList [3,5]) == Just (fromList (3:|[5]))

withNonEmpty

Arguments

:: r
value to return if set is empty

-> (NESet a -> r)
function to apply if set is not empty

-> Set a

-> r

O(log n). A general continuation-based way to consume a Set as if it were an NESet. withNonEmpty def f will take a Set. If set is empty, it will evaluate to def. Otherwise, a non-empty set NESet will be fed to the function f instead.

nonEmptySet == withNonEmpty Nothing Just

toSet :: NESet a -> Set a

O(log n). Convert a non-empty set back into a normal possibly-empty map, for usage with functions that expect Set.

Can be thought of as "obscuring" the non-emptiness of the set in its type. See the IsNotEmpty pattern.

nonEmptySet and maybe empty toSet form an isomorphism: they are perfect structure-preserving inverses of eachother.

toSet (fromList ((3,"a") :| [(5,"b")])) == Data.Set.fromList [(3,"a"), (5,"b")]

singleton :: a -> NESet a

O(1). Create a singleton set.

fromList :: Ord a => NonEmpty a -> NESet a

O(n*log n). Create a set from a list of elements.

toList :: NESet a -> NonEmpty a

O(n). Convert the set to a non-empty list of elements.

size :: NESet a -> Int

O(1). The number of elements in the set. Guaranteed to be greater than zero.

union :: Ord a => NESet a -> NESet a -> NESet a

O(m*log(n/m + 1)), m <= n. The union of two sets, preferring the first set when equal elements are encountered.

unions :: (Foldable1 f, Ord a) => f (NESet a) -> NESet a

The union of a non-empty list of sets

foldr :: (a -> b -> b) -> b -> NESet a -> b

O(n). Fold the elements in the set using the given right-associative binary operator, such that foldr f z == foldr f z . toAscList.

For example,

elemsList set = foldr (:) [] set

foldl :: (a -> b -> a) -> a -> NESet b -> a

O(n). Fold the elements in the set using the given left-associative binary operator, such that foldl f z == foldl f z . toAscList.

For example,

descElemsList set = foldl (flip (:)) [] set

foldr' :: (a -> b -> b) -> b -> NESet a -> b

O(n). A strict version of foldr. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

foldl' :: (a -> b -> a) -> a -> NESet b -> a

O(n). A strict version of foldl. Each application of the operator is evaluated before using the result in the next application. This function is strict in the starting value.

newtype MergeNESet a

Used for cartesianProduct

Constructors

MergeNESet

Fields

getMergeNESet :: NESet a

merge :: NESet a -> NESet a -> NESet a

Unsafely merge two disjoint sets. Only legal if all items in the first set are less than all items in the second set

valid :: Ord a => NESet a -> Bool

O(n). Test if the internal set structure is valid.

insertMinSet :: a -> Set a -> Set a

O(log n). Insert new value into a set where values are strictly greater than the new values That is, the new value must be strictly less than all values present in the Set. /The precondition is not checked./

While this has the same asymptotics as Data.Set.insert, it saves a constant factor for value comparison (so may be helpful if comparison is expensive) and also does not require an Ord instance for the value type.

insertMaxSet :: a -> Set a -> Set a

O(log n). Insert new value into a set where values are /strictly less than the new value. That is, the new value must be strictly greater than all values present in the Set. The precondition is not checked./

While this has the same asymptotics as Data.Set.insert, it saves a constant factor for value comparison (so may be helpful if comparison is expensive) and also does not require an Ord instance for the value type.

disjointSet :: Ord a => Set a -> Set a -> Bool

CPP for new functions not in old containers ---------------------------------------------

Comptability layer for disjoint.

powerSetSet :: Set a -> Set (Set a)

Comptability layer for powerSet.

disjointUnionSet :: Set a -> Set b -> Set (Either a b)

Comptability layer for disjointUnion.

cartesianProductSet :: Set a -> Set b -> Set (a, b)

Comptability layer for cartesianProduct.